requirements.txt
keyboard
flask
opencv-python
pygame
gym
gym[box2d]
Pillow
rich
gymnasium[box2d]
pyglet